Fix PDF's with uppercase file extension (#23)
* Fix PDF's with uppercase file extension

This is a quick fix for PDF's with upper case or mixed case extensions. But detection if a file is a PDF should be better performed on the file data. PDF's mostly start with `%PDF`, but why not use a standard solution as PDFJS or allready used gm to load the pdf and then see if it throws an error or not.